loginController

+++ ccElectricRent/CE.Data/Entity/Order.cs
@@ -8,7 +8,7 @@ namespace CE.Data.Entity
    public class Order
     {
         public int OrderId { set; get; }
-        public decimal TotalPrice { get; set; }
+        public float TotalPrice { get; set; }
         public DateTime CreateDate { set; get; }
         public Guid UserId { set; get; }
         public int VoucherId { get; set; }
